feat(tauri) core template isolated
