+++ lib/rust/Cargo.toml
@@ -31,11 +31,14 @@ uuid = { version = "0.7", features = ["v4"] }
 lazy_static = "1.3.0"
 includedir = "0.5.0"
 tiny_http = "0.6"
+clap = {version = "2.33", features = ["yaml"]}
 
 [build-dependencies]
 includedir_codegen = "0.5.0"
 
 [features]
+dev = []
+embedded-server = []
 all-api = []
 readTextFile = []
 readBinaryFile = []
