+++ lib/rust/build.rs
@@ -1,10 +1,15 @@
 extern crate includedir_codegen;
 
 use includedir_codegen::Compression;
+use std::env;
+
+static CARGOENV: &str = "cargo:rustc-env=";
 
 fn main() {
+  let dist_path = format!("{}/../../../../{}", env::var("OUT_DIR").unwrap(), "compiled-web");
+  println!("{}TAURI_DIST_DIR={}", CARGOENV, dist_path);
   includedir_codegen::start("ASSETS")
-    .dir("./target/compiled-web", Compression::Gzip)
+    .dir(dist_path, Compression::None)
     .build("data.rs")
     .unwrap();
 }
